requirements.txt
google-generativeai
streamlit
python-dotenv
st-chat-message